[Add] Checkbox선택 시 fetch 조건 console 확인
