+++ src/pages/Recommend/Recommend.js
@@ -1,6 +1,8 @@
 import React, { Component } from 'react';
 import { Link } from 'react-router-dom';
 
+import { GET_PRODUCTS_API } from '../../config';
+
 import { SurveyWelcome } from './Survey/SurveyWelcome';
 import { Survey01 } from './Survey/Survey01';
 import { Survey02 } from './Survey/Survey02';
@@ -13,8 +15,14 @@ export class Recommend extends Component {
   constructor() {
     super();
     this.state = {
-      surveyId: 1,
+      surveyId: 5,
       answer: [],
+      filterState: {
+        bone: false,
+        hair: false,
+        growth: false,
+        skin: false,
+      },
     };
   }
 
@@ -42,6 +50,54 @@ export class Recommend extends Component {
 
   handleNextButton = id => {
     this.setState({ surveyId: id + 1 });
+  };
+
+  makeCondition = () => {
+    const filterMatch = {
+      bone: 1,
+      hair: 2,
+      growth: 3,
+      skin: 4,
+    };
+
+    const filtered = Object.entries(this.state.filterState).reduce(
+      (acc, [key, value]) => {
+        if (!acc && value) {
+          return acc + `efficacy=${filterMatch[key]}`;
+        }
+
+        if (value) {
+          return acc + `&efficacy=${filterMatch[key]}`;
+        }
+        return acc;
+      },
+      ''
+    );
+
+    fetch(`${GET_PRODUCTS_API}?${filtered}`)
+      .then(res => res.json())
+      .then(data => {
+        this.setState({
+          productCard: data.message,
+        });
+      });
+    console.log(`filtered>>>>>`, `${GET_PRODUCTS_API}?${filtered}`);
+  };
+
+  handleCheckBox = event => {
+    const checkBoxName = event.target.name;
+    const checkBoxNameState = !this.state.filterState[checkBoxName];
+    this.setState(
+      {
+        filterState: {
+          ...this.state.filterState,
+          [checkBoxName]: checkBoxNameState,
+        },
+      },
+      () => {
+        this.makeCondition();
+      }
+    );
   };
 
   render() {
