+++ src/pages/Recommend/Survey/Survey04.js
@@ -1,71 +1,10 @@
 import React, { Component } from 'react';
 
-import { GET_PRODUCTS_API } from '../../../config';
-
 import './Survey.scss';
 
 export class Survey04 extends Component {
-  constructor() {
-    super();
-    this.state = {
-      filterState: {
-        bone: false,
-        hair: false,
-        growth: false,
-        skin: false,
-      },
-    };
-  }
-
-  makeCondition = () => {
-    const filterMatch = {
-      bone: 1,
-      hair: 2,
-      growth: 3,
-      skin: 4,
-    };
-
-    const filtered = Object.entries(this.state.filterState).reduce(
-      (acc, [key, value]) => {
-        if (!acc && value) {
-          return acc + `efficacy=${filterMatch[key]}`;
-        }
-
-        if (value) {
-          return acc + `&efficacy=${filterMatch[key]}`;
-        }
-        return acc;
-      },
-      ''
-    );
-
-    fetch(`${GET_PRODUCTS_API}?${filtered}`)
-      .then(res => res.json())
-      .then(data => {
-        this.setState({
-          productCard: data.message,
-        });
-      });
-  };
-
-  handleCheckBox = event => {
-    const checkBoxName = event.target.name;
-    const checkBoxNameState = !this.state.filterState[checkBoxName];
-    this.setState(
-      {
-        filterState: {
-          ...this.state.filterState,
-          [checkBoxName]: checkBoxNameState,
-        },
-      },
-      () => {
-        this.makeCondition();
-      }
-    );
-  };
-
   render() {
-    const { surveyId, handlePrevSubmmit } = this.props;
+    const { surveyId, handlePrevSubmmit, handleCheckBox } = this.props;
     return (
       <div className="Survey survey04">
         <h2 className="surveyTitle">
@@ -81,11 +20,23 @@ export class Survey04 extends Component {
         <hr />
         <form className="surveyForm">
           <div className="checkbox">
-            <input id="1" name="bone" className="inputGender" type="checkbox" />
+            <input
+              id="1"
+              name="bone"
+              className="inputGender"
+              type="checkbox"
+              onChange={handleCheckBox}
+            />
             <label>뼈</label>
           </div>
           <div className="checkbox">
-            <input id="2" name="hair" className="inputGender" type="checkbox" />
+            <input
+              id="2"
+              name="hair"
+              className="inputGender"
+              type="checkbox"
+              onChange={handleCheckBox}
+            />
             <label>모발</label>
           </div>
           <div className="checkbox">
@@ -94,11 +45,18 @@ export class Survey04 extends Component {
               name="growth"
               className="inputGender"
               type="checkbox"
+              onChange={handleCheckBox}
             />
             <label>성장</label>
           </div>
           <div className="checkbox">
-            <input id="4" name="skin" className="inputGender" type="checkbox" />
+            <input
+              id="4"
+              name="skin"
+              className="inputGender"
+              type="checkbox"
+              onChange={handleCheckBox}
+            />
             <label>피부</label>
           </div>
           <div className="buttonBox">
@@ -108,7 +66,7 @@ export class Survey04 extends Component {
             >
               이전
             </button>
-            <button className="nextButton" onClick={this.handleCheckBox}>
+            <button className="nextButton" onClick={handleCheckBox}>
               제출
             </button>
           </div>
